requirements.txt
websockets
python-dotenv
pyyaml
fastapi
uvicorn
httpx
pydantic_extra_types
